Make download scripts of object detection and single stage detector idempotent

+++ object_detection/scripts/download_dataset.sh
@@ -2,32 +2,29 @@
 set -ex
 # Get COCO 2014 data sets
 
-cd ./object_detection
-
-rm -rf data/
-mkdir -p data/coco
-pushd data/coco
+mkdir -p ./object_detection/data/coco
+cd ./object_detection/data/coco
 
 apt-get update
 apt-get install -y curl unzip
 
+if [ ! -f coco_annotations_minival.tgz ]; then
 curl -O https://dl.fbaipublicfiles.com/detectron/coco/coco_annotations_minival.tgz
+fi
 tar xzf coco_annotations_minival.tgz
 
+if [ ! -f train2014.zip ]; then
 curl -O http://images.cocodataset.org/zips/train2014.zip
-unzip train2014.zip
+fi
+unzip -o train2014.zip
 
+if [ ! -f val2014.zip ]; then
 curl -O http://images.cocodataset.org/zips/val2014.zip
-unzip val2014.zip
+fi
+unzip -o val2014.zip
 
+if [ ! -f annotations_trainval2014.zip ]; then
 curl -O http://images.cocodataset.org/annotations/annotations_trainval2014.zip
-unzip annotations_trainval2014.zip
+fi
+unzip -o annotations_trainval2014.zip
 
-# TBD: MD5 verification
-# $md5sum *.zip *.tgz
-#f4bbac642086de4f52a3fdda2de5fa2c  annotations_trainval2017.zip
-#cced6f7f71b7629ddf16f17bbcfab6b2  train2017.zip
-#442b8da7639aecaf257c1dceb8ba8c80  val2017.zip
-#2d2b9d2283adb5e3b8d25eec88e65064  coco_annotations_minival.tgz
-
-popd
